Rewrite itty-compression to use generics
This is a substantial rewrite of most of itty-compression to make use
of generics on the compression middlewares.

This means that all middlewares now only accept Requests, Responses,
and ResponseInit (for options) args. As a result, many input type
related tests have been scrapped.

The middlewares no longer care about the intermediate Response's
Content-Type marking and try to compress its body with the chosen
method.