Bump dependencies & fix for CVE-2024-4068
This includes a fix for CVE-2024-4068, which affected braces < 3.0.3
(which globby transitively depends on via fast-glob and micromatch).
These upgrades mean braces 3.0.3 is now used under the hood.

Bumped @types/node to 20.14.11
Bumped @vitest/coverage-v8 to 1.6.0
Bumped globby to 14.0.2
Bumped prettier to 3.3.3
Bumped rimraf to 5.0.9
Bumped rollup to 4.18.1
Bumped tslib to 2.6.3
Bumped TypeScript to 5.5.3
Bumped typescript-eslint to 7.16.1