Migrate to ESLint v9 "flat configs"
This involved turning the .eslintrc file into an eslint.config.mjs
file.

A few issues were highlighted by ESLint with this new configuration
setup:
- compressor.ts: @typescript-eslint/ prefer-nullish-coalescing says to
use nullish coalescence instead of logical or.
- constants.ts: @typescript-eslint/ no-unnecessary-type-assertion says
to remove the 'as const' from the string literals here.
- wrappers.ts: @typescript-eslint/ no-unnecessary-condition says an
optional chain is not needed in that location.

(future changes like this won't necessarily be listed out)

A new tsconfig.build.json was created to be used in building the
project for distribution (see rollup.config.mjs).
Other types of builds will be directed to the "out" directory. These
are needed to allow linting by ESLint for the test files without
including the test files in the release distribution.

rollup.config.mjs was formatted with prettier.

ESLint v9 also required dev dependencies on:
- globals
- @eslint/js

Added @stylistic/eslint-plugin to replace deprecated ESLint rules
such as "linebreak-style".

Bumped eslint to 9.17.0
Bumped typescript-eslint to 8.19.0