Bump dependencies
This includes several security fixes:
- GHSA-67mh-4wv8-2f99 (affected esbuild <= 0.24.2)
- CVE-2025-30208 (affected vite >= 6.1.0, < 6.1.2)
- CVE-2025-31125 (affected vite >= 6.1.0, < 6.1.3)
- CVE-2025-31486 (affected vite >= 6.1.0, < 6.1.4)
- CVE-2025-32395 (affected vite >= 6.1.0, < 6.1.5)

These upgrades mean esbuild is now set to 0.25.2 and vite to 6.2.6.
Both of these are dependencies of vitest.

Bumped @eslint/js to 9.24.0
Bumped @rollup/plugin-node-resolve to 16.0.1
Bumped @rollup/plugin-typescript to 12.1.2
Bumped @stylistic/eslint-plugin to 4.2.0
Bumped @types/node to 22.14.1
Bumped @vitest/coverage-v8 to 3.1.1
Bumped eslint to 9.24.0
Bumped eslint-config-prettier to 10.1.2
Bumped globals to 15.15.0
Bumped globby to 14.1.0
Bumped is-ci to 4.1.0
Bumped prettier to 3.5.3
Bumped rimraf to 6.0.1
Bumped rollup to 4.40.0
Bumped TypeScript to 5.8.3
Bumped typescript-eslint to 8.29.1
Bumped vitest to 3.1.1