Override rollup-plugin-copy version of globby
While Dependabot Security Alerts recognize that the project itself
does not use a vulnerable version of glob any more, it still sends
notification emails linking to the fixed alerts. I figured that the
transitively used copy of glob v7 was the cause and while the
maintainer of rollup-plugin-copy has updated their dependencies, they
have (so far) not made an actual new release to npm.

itty-compression should not be exposed to this since the plugin is
only used to copy the LICENCE file into the build output.

Override based on this comment:
https://github.com/vladshcherbin/rollup-plugin-copy/issues/77#issuecomment-2307846444